refactor : 로또 생성기 model로 이동
