+++ src/main/java/lotto/model/lotto/LottoGenerator.java
@@ -1,0 +1,7 @@
+package lotto.model.lotto;
+
+import lotto.model.number.LottoNumbers;
+
+public interface LottoGenerator {
+    LottoNumbers generateLottoNumbers(int minimumNumber, int maximumNumber, int lottoLength);
+}
