+++ src/main/java/lotto/model/number/BonusNumber.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.number;
 
 public class BonusNumber extends LottoNumber {
