+++ src/main/java/lotto/model/number/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.number;
 
 public class LottoNumber {
