--- src/main/java/lotto/util/LottoUtil.java
@@ -1,7 +1,0 @@
-package lotto.util;
-
-import lotto.model.LottoNumbers;
-
-public interface LottoUtil {
-    LottoNumbers generateLottoNumbers(int minimumNumber, int maximumNumber);
-}
