+++ src/test/java/lotto/util/RandomLottoGeneratorTest.java
@@ -4,17 +4,19 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import lotto.model.LottoNumber;
-import lotto.model.LottoNumbers;
+import lotto.model.lotto.AutoLottoNumbersGenerator;
+import lotto.model.lotto.LottoGenerator;
+import lotto.model.number.LottoNumber;
+import lotto.model.number.LottoNumbers;
 import org.junit.jupiter.api.Test;
 
 public class RandomLottoGeneratorTest {
 
-    LottoUtil randomLottoGenerator = new LottoNumbersGenerator();
+    LottoGenerator autoLottoNumbersGenerator = new AutoLottoNumbersGenerator();
 
     @Test
     void 중복_테스트() {
-        LottoNumbers numbers = randomLottoGenerator.generateLottoNumbers(1,45);
+        LottoNumbers numbers = autoLottoNumbersGenerator.generateLottoNumbers(1, 45, 6);
         List<Integer> actual = numbers.getLottoNumbers().stream()
                 .map(LottoNumber::getLottoNumber)
                 .distinct()
@@ -24,7 +26,7 @@ public class RandomLottoGeneratorTest {
 
     @Test
     void 범위_테스트() {
-        LottoNumbers numbers = randomLottoGenerator.generateLottoNumbers(1,45);
+        LottoNumbers numbers = autoLottoNumbersGenerator.generateLottoNumbers(1, 45, 6);
         numbers.getLottoNumbers().forEach(number -> {
             assertThat(number.getLottoNumber()).isGreaterThan(0).isLessThan(46);
         });
